+++ src/modules/contactPages/services/DeleteContactPagePostService.ts
@@ -14,11 +14,19 @@ class DeleteContactPagePostService {
     private userContactPagesRepository: IUserContactPagesRepository,
   ) {}
 
-  public async execute(id: string): Promise<void> {
+  public async execute(id: string, user_id: string): Promise<void> {
     const post = await this.contactPagePostsRepository.findById(id);
 
     if (!post) {
       throw new AppError('Contact page post not found!');
+    }
+
+    const contactPage = await this.userContactPagesRepository.findById(
+      post.contact_page_id,
+    );
+
+    if (!contactPage || contactPage.user_id !== user_id) {
+      throw new AppError('Contact page not found!');
     }
 
     await this.contactPagePostsRepository.delete(id);
